requirements.txt
dj-database-url==0.5.0
Django==2.1
django-heroku==0.3.1
gunicorn==20.0.4
h5py==2.10.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
numpy==1.19.2
opencv-python==4.4.0.44
scipy==1.4.1
tensorflow-cpu==2.2.0
whitenoise==5.2.0


